Add development release creation handling. Development builds on stored by manifest repository name.

+++ scripts/container_scripts/create-artifact.sh
@@ -1,22 +1,26 @@
 #!/bin/bash
 #=============================================================================================================================================================================
 #title: create-artifact.sh
-#description: Creates output release artifact tarball from built source.
+#description:  Creates output release artifact tarball from built source. 
+#              Non-development builds are stored in $CI_DIR/builds/<versioned-build>/release/.
+#              Development builds are stored in $CI_DIR/builds/development/release/<triggering-manifest-name>/.
 #=============================================================================================================================================================================
 set -e
 FULL_RELEASE_NAME="${RELEASE_NAME}_v${BUILD_VERSION}"
-RELEASE_DIR="/home/dev/Qualcomm/release/${FULL_RELEASE_NAME}"
+
+if [ "$BUILD_VERSION" == "development" ]; then
+    REPO_NAME=$(basename "$MANIFEST_REPOSITORY")
+    RELEASE_PATH="/home/dev/Qualcomm/release/${REPO_NAME}"
+else
+    RELEASE_PATH="/home/dev/Qualcomm/release"
+fi
+
+RELEASE_DIR="${RELEASE_PATH}/${FULL_RELEASE_NAME}"
 OUTPUT_TAR="${RELEASE_DIR}.tar.gz"
 
-#if release directory already exists, remove it
-if [ -d "$RELEASE_DIR" ]; then
-    rm -rf "$RELEASE_DIR"
-fi
-
-#if previous prebuilt tarball exists, remove it
-if [ -f "$OUTPUT_TAR" ]; then
-    rm -f "$OUTPUT_TAR"
-fi
+#if present, remove previous release directory and tarball
+if [ -d "$RELEASE_DIR" ]; then rm -rf "$RELEASE_DIR"; fi
+if [ -f "$OUTPUT_TAR" ]; then rm -f "$OUTPUT_TAR"; fi
 
 #create new prebuilt release
 cd /home/dev/build_scripts
@@ -26,10 +30,10 @@ python3 create_release.py \
     -b "${QCOM_ROOT_DIR}"
 
 #create fresh prebuilt tarball 
-cd /home/dev/Qualcomm/release
+cd "$RELEASE_PATH"
 tar -czvf "$OUTPUT_TAR" "$FULL_RELEASE_NAME"
 
-#verify tarball creation and remove non-compresssed release directory
+#verify tarball creation and remove un-compresssed release directory
 if [ -f "$OUTPUT_TAR" ]; then
     echo "Release artifact compressed at $OUTPUT_TAR"
     echo "Removing uncompressed release directory..."
